requirements.txt
mcp==1.27.0
openai==2.31.0
openai-whisper==20250625
python-dotenv==1.2.2
requests==2.33.1
python-telegram-bot==22.7
yt-dlp==2026.3.17
redis==7.4.0
qdrant-client==1.17.1
